Mock out firebase
