+++ runs/graphql/src/resolvers/me.test.js
@@ -1,3 +1,5 @@
+jest.mock("../firebase.js");
+import { getUser } from "../firebase.js";
 import { me } from "./me.js";
 
 test("me to be defined", () => {
@@ -5,9 +7,16 @@ test("me to be defined", () => {
 });
 
 test("me to return", async () => {
-  const user = await me();
-  expect(user).toEqual({
+  const fakeUser = {
     name: "Ada Lovelace",
     githubId: "adalovelace"
-  });
+  };
+  getUser.mockReturnValue(
+    Promise.resolve({
+      exists: true,
+      data: () => fakeUser
+    })
+  );
+  const user = await me();
+  expect(user).toEqual(fakeUser);
 });
